+++ app/jobs/vehicles/record_fitment_analytic_job.rb
@@ -5,11 +5,13 @@ module Vehicles
   class RecordFitmentAnalyticJob < ApplicationJob
     queue_as :low_priority
 
-    def perform(shop_id:, metric:, dimension: "all", day: Time.zone.today.to_s)
+    def perform(shop_id:, metric:, dimension: FitmentAnalytic::DIMENSION_ALL,
+                dimension_value: nil, day: Time.zone.today.to_s)
       shop = Shop.find_by(id: shop_id)
       return unless shop
 
-      FitmentAnalytic.increment(shop, metric: metric, dimension: dimension, day: Date.parse(day))
+      FitmentAnalytic.increment(shop, metric: metric, dimension: dimension,
+                                      dimension_value: dimension_value, day: Date.parse(day))
     end
   end
 end